// src/views/Signup/Nontsri/interfaces/student.ts
export interface Student {
  advisorId: string
  advisorNameEn: string
  advisorNameTh: string
  campusCode: string
  campusNameEn: string
  campusNameTh: string
  copenId: string
  copenNameEn: string
  copenNameTh: string
  departmentCode: string
  departmentNameEn: string
  departmentNameTh: string
  edulevelCode: string
  edulevelNameEn: string
  edulevelNameTh: string
  email: string
  facultyCode: string
  facultyNameEn: string
  facultyNameTh: string
  firstNameEn: string
  firstNameTh: string
  lastNameEn: string
  lastNameTh: string
  loginName: string
  majorCode: string
  majorNameEn: string
  majorNameTh: string
  middleNameEn: string | null
  middleNameTh: string | null
  mobileNo: string
  nationCode: string
  nationalityNameEn: string
  nationalityNameTh: string
  positionTh: string
  stdCode: string
  stdId: string
  studentStatusCode: string
  studentStatusNameEn: string
  studentStatusNameTh: string
  studentTypeCode: string
  studentTypeNameEn: string
  studentTypeNameTh: string
  studentYear: string
  titleEn: string
  titleTh: string
}